chore: revamped project architecture
