+++ src/io.rs
@@ -1,0 +1,4 @@
+//! Handles parsing and writing of the input and output files.
+
+pub mod parser;
+pub mod writer;
