+++ src/io/parser.rs
@@ -1,0 +1,2 @@
+//! Custom parser for consuming binaries and generating [`ir::IR`] objects.
+//! Directly handles the parsing of the input binary file and generation of the IR.
