+++ src/io/writer.rs
@@ -1,0 +1,1 @@
+//! Serializes [`ir::IR`] objects into binary files.
