+++ src/ir/auxdata.rs
