+++ src/ir/byte_interval.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 
-use crate::block::Block;
-use crate::symbolic_expression::SymbolicExpression;
+use super::block::Block;
+use super::symbolic_expression::SymbolicExpression;
 
 use node_derive::Node;
