+++ src/lib.rs
@@ -1,13 +1,7 @@
 //! This crate provides a Rust API for the GTIRB intermediate representation.
 //! It is generated from the GTIRB protobuf schema.
 
-pub mod auxdata;
-pub mod block;
-pub mod byte_interval;
-pub mod cfg;
+pub mod io;
 pub mod ir;
-pub mod module;
-pub mod node;
-pub mod section;
-pub mod symbol;
-pub mod symbolic_expression;
+pub mod prelude;
+pub mod utils;
