+++ src/prelude.rs
@@ -1,0 +1,3 @@
+//! Provides common imports and re-exports for all modules.
+
+pub use crate::ir::*;
