+++ src/utils.rs
@@ -1,0 +1,3 @@
+//! Shared utilities used across the project.
+
+pub mod error;
