+++ src/utils/error.rs
@@ -1,0 +1,1 @@
+//! Custom error types implementing the `std::error::Error` trait for robust error handling.
